--- moqt-server-v2/src/modules/session_handlers/message_controller_trait.rs
@@ -1,1 +1,0 @@
-pub(crate) trait MessageControllerTrait {}
