+++ moqt-server-v2/src/modules/session_handlers/moqt_connection_creator.rs
@@ -1,5 +1,5 @@
-use crate::modules::session_handlers::transport_connection_creator::TransportConnectionCreator;
 use crate::modules::session_handlers::moqt_connection::MOQTConnection;
+use crate::modules::session_handlers::transport_connection_creator::TransportConnectionCreator;
 
 pub(crate) struct MOQTConnectionCreator {
     transport_creator: Box<dyn TransportConnectionCreator>,
