--- moqt-server-v2/src/modules/session_handlers/moqt_request_id_manager.rs
@@ -1,9 +1,0 @@
-pub(crate) struct MOQTRequestIdManager {
-    request_id: std::sync::atomic::AtomicU64
-}
-
-impl MOQTRequestIdManager {
-    pub(crate) fn get_next_request_id(&mut self) -> u64 {
-        self.request_id.load(std::sync::atomic::Ordering::Relaxed)
-    }
-}
